pyproject.toml
[tool.poetry]
name = "prognet_messages"
version = "0.1.0"
description = ""
authors = ["Timofey Danshin <t_danshin@gcore.lu>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.6.2"
click = "^7.1.2"
pyyaml = "^5.3.1"
aio_pika = "^6.7.0"
aiomysql = "^0.0.20"
pyjwt = "^1.7.1"
aiohttp-cors = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
